pkgs/standards/peagen/pyproject.toml
[project]
name = "peagen"
version = "0.1.4.dev3"
description = "Swarmauri's Peagan - An AI-driven contextual, dependency-based scaffolding tool for rapid content generation."
authors = [{ name = "Jacob Stewart", email = "jacob@swarmauri.com" }]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/swarmauri/swarmauri-sdk/pkgs/standards/peagen/peagen"
requires-python = ">=3.10,<3.13"
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [

    # are servers optional?
    "httpx[http2]>=0.27", # required
    "websockets>=12", # need to push out to a standalone with extra
    "h11>=0.16.0",
    
    # are servers optional?
    "uvicorn[standard]>=0.29",
    "fastapi>=0.111",



    "python-dotenv", # does this come with standard?
    "pyyaml", # should come with swarmauri_standard, but we can explicitly mention here

    "jsonpatch>=1.33",
    "jsonschema>=4.18.5",

    "jinja2>=3.1.6", # should come with j2prompttemplate, but is it required anywhere else on its own?
    "inflect", # should come with j2prompttemplate, but is it required anywhere else on its own?
    "pgpy>=0.6.0",
    "psutil>=6.0",  # required for IO evaluator

    "typer",   # for cli
    "colorama",# for cli
    
    "textual", # for tui
    "rich", # for tui

    "swarmauri_core",
    "swarmauri_base",
    "swarmauri_standard",
    "swarmauri",
    "swarmauri_prompt_j2prompttemplate",

    "pydantic-settings>=2.2", 
    "pydantic[email]>=2.7", # why is this necessary?

    # plugins —————————————————————————————————————————————————————————————————————————
    # result_backends/indexers
    "sqlalchemy>=2.0",
    "alembic",
    "asyncpg>=0.30.0", # need to push out to a standalone with extra
    "psycopg2-binary", # need to push out to a standalone with extra
    "aiosqlite>=0.19.0", # needed for the SQLite fallback result backend
    
    # brokers (publishers/consumer dependencies)
    "redis", # this publisher needs to be pushed to its own standalone
    "redis[hiredis]>=5.0", # provides redis.asyncio
    "pika", # need to push out to a standalone with extra

    # artifacts
    "PyGithub", # this storage adapter needs to be pushed to its own standalone
    "minio", # this storage adapter needs to be pushed to its own standalone
    "GitPython",
    "paramiko",
    "pygit2",
    "s3fs",
    "pytest-monitor>=1.6.6",
]


[tool.uv.sources]
swarmauri_core = { workspace = true }
swarmauri_base = { workspace = true }
swarmauri_standard = { workspace = true }

[tool.pytest.ini_options]
norecursedirs = ["combined", "scripts"]
markers = [
    "test: standard test",
    "unit: Unit tests",
    "i9n: Integration tests",
    "r8n: Regression tests",
    "timeout: mark test to timeout after X seconds",
    "xpass: Expected passes",
    "xfail: Expected failures",
    "acceptance: Acceptance tests",
    "perf: Performance tests that measure execution time and resource usage",
]
timeout = 300
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)s] %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
asyncio_default_fixture_loop_scope = "function"

[project.scripts]
peagen = "peagen.cli:app"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-xdist>=3.6.1",
    "pytest-json-report>=1.5.0",
    "python-dotenv",
    "httpx>=0.27.0",
    "flake8>=7.0",
    "pytest-timeout>=2.3.1",
    "ruff>=0.9.9",
    "pytest-benchmark>=4.0.0",
    "textual>=3.3.0",
]

[project.entry-points."peagen.template_sets"]
react_atom                            = "peagen.template_sets.react_atom"
svelte_atom                           = "peagen.template_sets.svelte_atom"
vue_atom                              = "peagen.template_sets.vue_atom"
swarmauri_base                        = "peagen.template_sets.swarmauri_base"
swarmauri_core                        = "peagen.template_sets.swarmauri_core"
swarmauri_standard                    = "peagen.template_sets.swarmauri_standard"
swarmauri_standard_standalone         = "peagen.template_sets.swarmauri_standard_standalone"
swarmauri_community                   = "peagen.template_sets.swarmauri_community"
python_orm                            = "peagen.template_sets.python_orm"
rust_python_pkg                       = "peagen.template_sets.rust_python_pkg"
cpp_python_pkg                        = "peagen.template_sets.cpp_python_pkg"
init-project                          = "peagen.template_scaffold.project"
init-template-set                     = "peagen.template_scaffold.template_set"
init-doe-spec                         = "peagen.template_scaffold.doe_spec"
init-ci                               = "peagen.template_scaffold.ci"

[project.entry-points."peagen.plugins.storage_adapters"]
minio   = "peagen.plugins.storage_adapters.minio_storage_adapter:MinioStorageAdapter"
file    = "peagen.plugins.storage_adapters.file_storage_adapter:FileStorageAdapter"
github    = "peagen.plugins.storage_adapters.github_storage_adapter:GithubStorageAdapter"
gh_release    = "peagen.plugins.storage_adapters.gh_release_storage_adapter:GithubReleaseStorageAdapter"

[project.entry-points."peagen.plugins.git_filters"]
minio   = "peagen.plugins.git_filters.minio_filter:MinioFilter"
gh_release    = "peagen.plugins.git_filters.gh_release_filter:GithubReleaseFilter"
s3fs    = "peagen.plugins.git_filters.s3fs_filter:S3FSFilter"

[project.entry-points."peagen.plugins.vcs"]
git = "peagen.plugins.vcs.git_vcs:GitVCS"

[project.entry-points."peagen.plugins.publishers"]
redis   = "peagen.plugins.publishers.redis_publisher:RedisPublisher"
webhook = "peagen.plugins.publishers.webhook_publisher:WebhookPublisher"
rabbitmq = "peagen.plugins.publishers.rabbitmq_publisher:RabbitMQPublisher"

[project.entry-points."peagen.plugins.result_backends"]
local_fs = "peagen.plugins.result_backends.localfs_backend:LocalFsResultBackend"
postgres = "peagen.plugins.result_backends.postgres_backend:PostgresResultBackend"
in_memory = "peagen.plugins.result_backends.in_memory_backend:InMemoryResultBackend"

[project.entry-points."peagen.plugins.queues"]
redis = "peagen.plugins.queues.redis_queue:RedisQueue"
in_memory = "peagen.plugins.queues.in_memory_queue:InMemoryQueue"

[project.entry-points."peagen.plugins.selectors"]
result_backend = "peagen.plugins.selectors.result_backend_selector:ResultBackendSelector"
bootstrap = "peagen.plugins.selectors.bootstrap_selector:BootstrapSelector"
input = "peagen.plugins.selectors.input_selector:InputSelector"

[project.entry-points."peagen.plugins.mutators"]
default_mutator = "peagen.plugins.mutators.default_mutator:DefaultMutator"
echo_mutator = "peagen.plugins.mutators.echo_mutator:EchoMutator"
llm_prompt = "peagen.plugins.mutators.llm_prompt:LlmRewrite"
llm_prog_rewrite = "peagen.plugins.mutators.llm_prog_rewrite:LlmProgRewrite"

[project.entry-points."peagen.evaluators"]
psutil_io = "peagen.evaluators.psutil_io:PsutilIOEvaluator"
perf_regression = "peagen.evaluators.pytest_perf_regression:PytestPerfRegressionEvaluator"
monitor = "peagen.evaluators.pytest_monitor:PytestMonitorEvaluator"
memray = "peagen.evaluators.pytest_memray_evaluator:PytestMemrayEvaluator"
pytest_profiling = "peagen.evaluators.pytest_profiling:PytestProfilingEvaluator"
benchmark = "peagen.evaluators.benchmark:PytestBenchmarkEvaluator"
simple_time = "peagen.evaluators.simple_time:SimpleTimeEvaluator"

[tool.setuptools.package-data]
"peagen.schemas" = ["*.json", "extras/*.json"]
"peagen" = [
  "tests/examples/locking_demo/*",
  "tests/examples/peagen_tomls/*",
  "tests/examples/2x2Grid/*",
  "alembic.ini",
  "migrations/**",
]
